requirements.bak.txt
gunicorn==23.0.0
alembic==1.13.3
APScheduler==3.10.4
SQLAlchemy==2.0.36
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
tabulate==0.9.0
python-dotenv==1.0.1
ultralytics==8.3.2